Updates, mostly just sync with others and kill all of the old home grown stuff.

+++ distrib/hp300/ramdisk/install.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-#	$OpenBSD: install.sh,v 1.3 1997/04/21 07:32:12 downsj Exp $
+#	$OpenBSD: install.sh,v 1.4 1997/10/13 07:47:22 downsj Exp $
 #	$NetBSD: install.sh,v 1.5.2.8 1996/08/27 18:15:05 gwr Exp $
 #
 # Copyright (c) 1996 The NetBSD Foundation, Inc.
@@ -44,7 +44,7 @@
 FILESYSTEMS="/tmp/filesystems"		# used thoughout
 FQDN=""					# domain name
 
-trap "umount /tmp >> /dev/null 2>&1" 0
+trap "umount /tmp > /dev/null 2>&1" 0
 
 MODE="install"
 
@@ -64,6 +64,7 @@ MODE="install"
 #	md_native_fstype()	- native filesystem type for disk installs
 #	md_native_fsopts()	- native filesystem options for disk installs
 #	md_makerootwritable()	- make root writable (at least /tmp)
+#	md_machine_arch()	- get machine architecture
 
 # include machine dependent subroutines
 . install.md
@@ -71,40 +72,37 @@ MODE="install"
 # include common subroutines
 . install.sub
 
-# Deal with terminal issues
-md_set_term
+# which sets?
+THESETS="$ALLSETS $MDSETS"
 
-# Good {morning,afternoon,evening,night}.
-md_welcome_banner
-echo -n "Proceed with installation? [n] "
+if [ "`df /`" = "`df /mnt`" ]; then
+	# Good {morning,afternoon,evening,night}.
+	md_welcome_banner
+	echo -n "Proceed with installation? [n] "
+else
+	echo "You seem to be trying to restart an interrupted installation!"
+	echo ""
+	echo "You can try to skip the disk preparation steps and continue,"
+	echo "otherwise you should reboot the miniroot and start over..."
+	echo -n "Skip disk initialization? [n] "
+fi
 getresp "n"
 case "$resp" in
 	y*|Y*)
+		echo	""
+		echo	"Cool!  Let's get to it..."
 		;;
 	*)
 		md_not_going_to_install
 		exit
 		;;
 esac
-
-echo -n "Is this a (s)napshot or (r)elease? [s] "
-getresp "s"
-case "$resp" in
-	s*|S*)
-		THESETS="$SNAPSETS $KERNSETS"
-		;;
-	r*|R*)
-		THESETS="$ALLSETS $KERNSETS"
-		;;
-	*)
-		md_not_going_to_install
-		exit
-		;;
-esac
-echo "Cool!  Let's get to it..."
 
 # XXX Work around vnode aliasing bug (thanks for the tip, Chris...)
-ls -l /dev >> /dev/null 2>&1
+ls -l /dev > /dev/null 2>&1
+
+# Deal with terminal issues
+md_set_term
 
 # Get timezone info
 get_timezone
@@ -112,25 +110,33 @@ get_timezone
 # Make sure we can write files (at least in /tmp)
 # This might make an MFS mount on /tmp, or it may
 # just re-mount the root with read-write enabled.
-md_makerootwritable
+if [ "`df /`" = "`df /tmp`" ]; then
+	md_makerootwritable
+fi
 
-# Install the shadowed disktab file; lets us write to it for temporary
-# purposes without mounting the miniroot read-write.
-cp /etc/disktab.shadow /tmp/disktab.shadow
+# Get the machine architecture (must be done after md_makerootwritable)
+ARCH=`md_machine_arch`
 
-while [ "X${ROOTDISK}" = "X" ]; do
-	getrootdisk
-done
+if [ "`df /`" = "`df /mnt`" ]; then
+	# Install the shadowed disktab file; lets us write to it for temporary
+	# purposes without mounting the miniroot read-write.
+	if [ -f /etc/disktab.shadow ]; then
+		cp /etc/disktab.shadow /tmp/disktab.shadow
+	fi
 
-# Deal with disklabels, including editing the root disklabel
-# and labeling additional disks.  This is machine-dependent since
-# some platforms may not be able to provide this functionality.
-md_prep_disklabel ${ROOTDISK}
+	while [ "X${ROOTDISK}" = "X" ]; do
+		getrootdisk
+	done
 
-# Assume partition 'a' of $ROOTDISK is for the root filesystem.  Loop and
-# get the rest.
-# XXX ASSUMES THAT THE USER DOESN'T PROVIDE BOGUS INPUT.
-cat << \__get_filesystems_1
+	# Deal with disklabels, including editing the root disklabel
+	# and labeling additional disks.  This is machine-dependent since
+	# some platforms may not be able to provide this functionality.
+	md_prep_disklabel ${ROOTDISK}
+
+	# Assume partition 'a' of $ROOTDISK is for the root filesystem.
+	# Loop and get the rest.
+	# XXX ASSUMES THAT THE USER DOESN'T PROVIDE BOGUS INPUT.
+	cat << \__get_filesystems_1
 
 You will now have the opportunity to enter filesystem information.
 You will be prompted for device name and mount point (full path,
@@ -143,70 +149,103 @@ to resolve any filesystem order dependencies.
 
 __get_filesystems_1
 
-echo	"The following will be used for the root filesystem:"
-echo	"	${ROOTDISK}a	/"
+	echo	"The following will be used for the root filesystem:"
+	echo	"	${ROOTDISK}a	/"
 
-echo	"${ROOTDISK}a	/" > ${FILESYSTEMS}
+	echo	"${ROOTDISK}a /" > ${FILESYSTEMS}
 
-resp="X"	# force at least one iteration
-while [ "X$resp" != X"done" ]; do
-	echo	""
-	echo -n	"Device name? [done] "
-	getresp "done"
-	case "$resp" in
-	done)
-		;;
+	resp="X"	# force at least one iteration
+	while [ "X$resp" != X"done" ]; do
+		echo	""
+		echo -n	"Device name? [done] "
+		getresp "done"
+		case "$resp" in
+		done)
+			;;
 
-	*)
-		_device_name=`basename $resp`
+		*)
+			_device_name=`basename $resp`
 
-		# force at least one iteration
-		_first_char="X"
-		while [ "X${_first_char}" != X"/" ]; do
-			echo -n "Mount point? "
-			getresp ""
-			_mount_point=$resp
-			if [ "X${_mount_point}" = X"/" ]; then
-				# Invalid response; no multiple roots
-				_first_char="X"
-			else
+			# force at least one iteration
+			_first_char="X"
+			while [ "X${_first_char}" != X"/" ]; do
+				echo -n "Mount point? "
+				getresp ""
+				_mount_point=$resp
 				_first_char=`firstchar ${_mount_point}`
+				if [ "X${_first_char}" != X"/" ]; then
+					echo "mount point must be an absolute path!"
+				fi
+			done
+			if [ "X${_mount_point}" = X"/" ]; then
+				echo "root mount point already taken care of!"
+			else
+				echo "${_device_name} ${_mount_point}" \
+					>> ${FILESYSTEMS}
 			fi
-		done
-		echo "${_device_name}	${_mount_point}" >> ${FILESYSTEMS}
-		resp="X"	# force loop to repeat
-		;;
-	esac
-done
-
-echo	""
-echo	"You have configured the following devices and mount points:"
-echo	""
-cat ${FILESYSTEMS}
-echo	""
-echo	"Filesystems will now be created on these devices.  If you made any"
-echo -n	"mistakes, you may edit this now.  Edit? [n] "
-getresp "n"
-case "$resp" in
-	y*|Y*)
-		${EDITOR} ${FILESYSTEMS}
-		;;
-	*)
-		;;
-esac
-
-# Loop though the file, place filesystems on each device.
-echo	"Creating filesystems..."
-(
-	while read _device_name _junk; do
-		newfs /dev/r${_device_name}
-		echo ""
+			resp="X"	# force loop to repeat
+			;;
+		esac
 	done
-) < ${FILESYSTEMS}
+
+	echo	""
+	echo	"You have configured the following devices and mount points:"
+	echo	""
+	cat ${FILESYSTEMS}
+	echo	""
+	echo	"Filesystems will now be created on these devices."
+	echo 	"If you made any mistakes, you may edit this now."
+	echo -n	"Edit using ${EDITOR}? [n] "
+	getresp "n"
+	case "$resp" in
+		y*|Y*)
+			${EDITOR} ${FILESYSTEMS}
+			;;
+		*)
+			;;
+	esac
+	echo 	""
+	echo	 "The next step will overwrite any existing data on:"
+	(
+		echo -n "	"
+		while read _device_name _junk; do
+			echo -n "${_device_name} "
+		done
+		echo ""
+	) < ${FILESYSTEMS}
+	echo	""
+
+	echo -n	"Are you really sure that you're ready to proceed? [n] "
+	getresp "n"
+	case "$resp" in
+		y*|Y*)
+			;;
+		*)
+			echo "ok, try again later..."
+			exit
+			;;
+	esac
+
+	# Loop though the file, place filesystems on each device.
+	echo	"Creating filesystems..."
+	(
+		while read _device_name _junk; do
+			newfs /dev/r${_device_name}
+			echo ""
+		done
+	) < ${FILESYSTEMS}
+else
+	# Get the root device
+	ROOTDISK=`df /mnt | sed -e '/^\//!d' -e 's/\/dev\/\([^ ]*\)[a-p] .*/\1/'`
+	while [ "X${ROOTDISK}" = "X" ]; do
+		getrootdisk
+	done
+fi
 
 # Get network configuration information, and store it for placement in the
 # root filesystem later.
 cat << \__network_config_1
+
 You will now be given the opportunity to configure the network.  This will
 be useful if you need to transfer the installation sets via FTP or NFS.
 Even if you choose not to transfer installation sets that way, this
@@ -224,17 +263,17 @@ case "$resp" in
 		if [ -f /etc/myname ]; then
 			resp=`cat /etc/myname`
 		fi
-		echo -n "Enter system hostname: [$resp] "
 		while [ "X${resp}" = X"" ]; do
+			echo -n "Enter system hostname: [$resp] "
 			getresp "$resp"
 		done
 		hostname $resp
 		echo $resp > /tmp/myname
 
-		echo -n "Enter DNS domain name: "
 		resp=""		# force at least one iteration
 		while [ "X${resp}" = X"" ]; do
-			getresp ""
+			echo -n "Enter DNS domain name: [$FQDN] "
+			getresp "$FQDN"
 		done
 		FQDN=$resp
 
@@ -243,8 +282,8 @@ case "$resp" in
 		echo -n "Enter IP address of default route: [none] "
 		getresp "none"
 		if [ "X${resp}" != X"none" ]; then
-			route delete default >> /dev/null 2>&1
-			if route add default $resp >> /dev/null ; then
+			route delete default > /dev/null 2>&1
+			if route add default $resp > /dev/null ; then
 				echo $resp > /tmp/mygate
 			fi
 		fi
@@ -275,8 +314,10 @@ case "$resp" in
 		cat /tmp/hosts
 		echo ""
 		echo "You may want to edit the host table in the event that"
-		echo "you need to mount an NFS server."
-		echo -n "Would you like to edit the host table? [n] "
+		echo "you are doing an NFS installation or an FTP installation"
+		echo "without a name server and want to refer to the server by"
+		echo "name rather than by its numeric ip address."
+		echo -n "Would you like to edit the host table with ${EDITOR}? [n] "
 		getresp "n"
 		case "$resp" in
 			y*|Y*)
@@ -318,22 +359,23 @@ __network_config_2
 		;;
 esac
 
-# Now that the network has been configured, it is safe to configure the
-# fstab.
-(
-	while read _dev _mp; do
-		if [ "$mp" = "/" ]; then
-			echo /dev/$_dev $_mp ffs rw 1 1
-		else
-			echo /dev/$_dev $_mp ffs rw 1 2
-		fi
-	done
-) < ${FILESYSTEMS} > /tmp/fstab
+if [ "`df /`" = "`df /mnt`" ]; then
+	# Now that the network has been configured, it is safe to configure the
+	# fstab.
+	(
+		while read _dev _mp; do
+			if [ "$_mp" = "/" ]; then
+				echo /dev/$_dev $_mp ffs rw 1 1
+			else
+				echo /dev/$_dev $_mp ffs rw 1 2
+			fi
+		done
+	) < ${FILESYSTEMS} > /tmp/fstab
 
-echo	"The fstab is configured as follows:"
-echo	""
-cat /tmp/fstab
-cat << \__fstab_config_1
+	echo	"The fstab is configured as follows:"
+	echo	""
+	cat /tmp/fstab
+	cat << \__fstab_config_1
 
 You may wish to edit the fstab.  For example, you may need to resolve
 dependencies in the order which the filesystems are mounted.  You may
@@ -342,20 +384,21 @@ This would be especially useful if you plan to keep '/usr' on an NFS
 server.
 
 __fstab_config_1
-echo -n	"Edit the fstab? [n] "
-getresp "n"
-case "$resp" in
-	y*|Y*)
-		${EDITOR} /tmp/fstab
-		;;
+	echo -n	"Edit the fstab with ${EDITOR}? [n] "
+	getresp "n"
+	case "$resp" in
+		y*|Y*)
+			${EDITOR} /tmp/fstab
+			;;
 
-	*)
-		;;
-esac
+		*)
+			;;
+	esac
 
-echo ""
-munge_fstab /tmp/fstab /tmp/fstab.shadow
-mount_fs /tmp/fstab.shadow
+	echo ""
+	munge_fstab /tmp/fstab /tmp/fstab.shadow
+	mount_fs /tmp/fstab.shadow
+fi
 
 mount | while read line; do
 	set -- $line
@@ -373,46 +416,60 @@ mount | while read line; do
 	fi
 done
 
-install_sets $ALLSETS $MDSETS
+install_sets $THESETS
 
 # Copy in configuration information and make devices in target root.
-(
-	cd /tmp
-	for file in fstab hostname.* hosts myname mygate resolv.conf; do
-		if [ -f $file ]; then
-			echo -n "Copying $file..."
-			cp $file /mnt/etc/$file
-			echo "done."
-		fi
-	done
 
-	# If no zoneinfo on the installfs, give them a second chance
-	if [ ! -e /usr/share/zoneinfo ]; then
-		get_timezone
-	fi
-	if [ ! -e /mnt/usr/share/zoneinfo ]; then
-		echo "Cannot install timezone link..."
-	else
-		echo -n "Installing timezone link..."
-		rm -f /mnt/etc/localtime
-		ln -s /usr/share/zoneinfo/$TZ /mnt/etc/localtime
+if [ ! -d /mnt/etc -o ! -d /mnt/usr/share/zoneinfo -o ! -d /mnt/dev ]; then
+	echo "Something needed to complete the installation seems"
+	echo "to be missing, did you forget to extract a required set?"
+	echo ""
+	echo "Please review the installation notes and try again..."
+	echo ""
+	echo "You *may* be able to correct the problem and type 'install'"
+	echo "without having to extract all of the distribution sets again."
+	exit
+fi
+
+cd /tmp
+for file in fstab hostname.* hosts myname mygate resolv.conf; do
+	if [ -f $file ]; then
+		echo -n "Copying $file..."
+		cp $file /mnt/etc/$file
 		echo "done."
 	fi
-	if [ ! -x /mnt/dev/MAKEDEV ]; then
-		echo "No /dev/MAKEDEV installed, something is wrong here..."
-	else
-		# Check if a device exists... snapshots have a dev.tar.gz
-		if [ ! -e /mnt/dev/rsd0a ]; then
-			echo -n "Making devices..."
-			cd /mnt/dev
-			sh MAKEDEV all
-			echo "done."
-		fi
-	fi
-	md_copy_kernel
+done
 
-	md_installboot ${ROOTDISK}
-)
+# If no zoneinfo on the installfs, give them a second chance
+if [ ! -e /usr/share/zoneinfo ]; then
+	get_timezone
+fi
+if [ ! -e /mnt/usr/share/zoneinfo ]; then
+	echo "Cannot install timezone link..."
+else
+	echo -n "Installing timezone link..."
+	rm -f /mnt/etc/localtime
+	ln -s /usr/share/zoneinfo/$TZ /mnt/etc/localtime
+	echo "done."
+fi
+
+
+md_copy_kernel
+
+md_installboot ${ROOTDISK}
+
+if [ ! -x /mnt/dev/MAKEDEV ]; then
+	echo "No /dev/MAKEDEV installed, something is wrong here..."
+	exit
+fi
+
+echo -n "Making all devices..."
+#pid=`twiddle`
+cd /mnt/dev
+sh MAKEDEV all
+#kill $pid
+echo "done."
+cd /
 
 unmount_fs /tmp/fstab.shadow
